extract racer.go
